% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_some_fonts.R
\name{download_some_fonts}
\alias{download_some_fonts}
\title{Download Some Fonts from Google}
\usage{
download_some_fonts(unzip = TRUE)
}
\arguments{
\item{unzip}{logical, defaults to TRUE If TRUE, the function unzips the contents
of the file it downloaded into the directory in which you downloaded the zip
file.}
}
\value{
\code{download_some_fonts()} downloads some fonts from Google, if
they don't already exist in the package's directory, and optionally unzips
them. It ends with a message about where these fonts were downloaded to
assist the user in locating these fonts (so that they may install them on
their operating system).
}
\description{
\code{download_some_fonts()} is a simple function that will
download some fonts from Google and stick them in the package directory.
}
\details{
\code{download_some_fonts()} assumes an active internet connection
in order to download some fonts from Google. It also assumes you don't have
these fonts yet, or do not have them installed. Loading the package will
advise you what fonts it suggests you have installed, pointing you in the
direction of what you may want to install. At worst, this function downloads
stuff you already have installed or downloaded in another directory on your
computer. You could always go right to Google and do this manually.

Right now, the function won't download a zip file for the font if it detects
the zip file it would download is already in the directory.
}
\examples{

\dontrun{
# Here's where the data are going to be downloaded.
system.file("fonts", package="stevethemes")
# Now, let's download the data.
download_some_fonts()
}
}
\author{
Steven V. Miller
}
