% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example_plot}
\alias{example_plot}
\title{Create an example plot (in \pkg{ggplot2}) for experimenting with different styles}
\usage{
example_plot(type = "bar")
}
\arguments{
\item{type}{a type}
}
\value{
\code{example_plot()} returns a plot made in \pkg{ggplot2}, allowing
the user to experiment with different themes to their hearts content.
}
\description{
\code{example_plot()} allows you to experiment with some
\pkg{ggplot2} themes, like those in this package, by seeing them applied to
some example plots.
}
\details{
I'll add more here in a little bit.
}
\examples{

library(ggplot2)

example_plot()

example_plot(type = "faceted_line")

}
\author{
Steven V. Miller
}
