% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_c.R
\name{g_c}
\alias{g_c}
\title{Specify a Custom Color from \code{steve_hex}}
\usage{
g_c(x)
}
\arguments{
\item{x}{a label coinciding with a value in the \code{color} column of the
\code{steve_hex} data frame}
}
\value{
\code{g_c()} returns a character vector, in particular, a hex triplet
that coincides with the label supplied in the function. Use it for specifying
a custom fill or color in a plot.
}
\description{
\code{g_c()} is a simple function that interfaces with the
\code{steve_hex} data frame to return a hex value associated with a named
color.
}
\details{
Check \code{steve_hex} in this same package for the custom colors
included
}
\examples{

g_c("su_blue")

g_c("martel_pink")
}
\author{
Steven V. Miller
}
