% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/how_to_install_fonts.R
\name{how_to_install_fonts}
\alias{how_to_install_fonts}
\title{Learn how to install fonts on your computer}
\usage{
how_to_install_fonts(os = "any")
}
\arguments{
\item{os}{a character vector coinciding with the information you want, contingent
on your operating system. If \code{os} is "any" (the default), it returns a
generic message about how to install fonts. If \code{os} is "windows", "mac",
or "linux", you get more information about font installation on your operating
system.}
}
\value{
\code{how_to_install_fonts()} returns a message communicating how to
install fonts on the user's operating system.
}
\description{
\code{how_to_install_fonts()} communicates to the user how they
can install fonts on their computer.
}
\details{
The 'any' option for the \code{os} argument is necessarily less
informative. If you know whether you're using Windows, a Mac, or Linux, you
should change the \code{os} argument.
}
\examples{

how_to_install_fonts("windows")

how_to_install_fonts("mac")
}
\author{
Steven V. Miller
}
