% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{martel_ratings}
\alias{martel_ratings}
\title{CAGEMATCH Ratings of Rick Martel}
\format{
A data frame the following 2 variables.
\describe{
\item{\code{date}}{a date vector for when the rating was posted on CAGEMATCH}
\item{\code{value}}{an individual person's rating of Rick Martel (on a 1-10 scale)}
}
}
\usage{
martel_ratings
}
\description{
This is a simple data frame of ratings (on a 1-10 scale) of Rick Martel.
}
\details{
Data exist to be used an example bar chart. CAGEMATCH (stylized in
all caps) is an internet wrestling database, for which these ratings are fan
submissions. Rick Martel is an objective 10/10.
}
\keyword{datasets}
