% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{se_counties_gdppc}
\alias{se_counties_gdppc}
\title{GDP per Capita of Swedish Counties, 2001-2020}
\format{
A data frame the following 4 variables.
\describe{
\item{\code{nuts}}{the Nomenclature of Territorial Units for Statistics (NUTS) code for the county}
\item{\code{county}}{the name of the county}
\item{\code{year}}{the year of observation}
\item{\code{value}}{the GDP per capita of the county in nominal SEK}
}
}
\usage{
se_counties_gdppc
}
\description{
This is a simple data frame of the GDP per capita of Swedish counties from
2001 to 2020.
}
\details{
Data come from the OECD and exist to be used as a faceted line chart.
}
\keyword{datasets}
