% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{steve_hex}
\alias{steve_hex}
\title{Some Hex Triplets I Find Useful/Interesting/Fun}
\format{
A data frame the following 2 variables.
\describe{
\item{\code{color}}{a character vector describing the color in question}
\item{\code{hex}}{a hex triplet (with preceding hashtag) of the color}
}
}
\usage{
steve_hex
}
\description{
This is a simple data frame with labels corresponding to hex triplets (i.e.
web colors) that I find useful, interesting, or fun.
}
\details{
Data exist to be used by \code{g_c()} in this same package.
}
\keyword{datasets}
