% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_steve.R
\name{theme_steve}
\alias{theme_steve}
\title{Steve's Preferred \pkg{ggplot2} Themes and Assorted Stuff}
\usage{
theme_steve(style = "web", ...)
}
\arguments{
\item{style}{various styles/adjustments to make to the base theme}

\item{...}{optional stuff, called for side effects}
}
\description{
\code{theme_steve()} is my default theme framework for graphs I
make with \pkg{ggplot2}. It starts \code{theme_bw()}, which is available in
\pkg{ggplot2}, but adjusts the margins and axes a bit to my liking. The end
result is, I think, a lovely template for graphs I make in R.
}
\details{
I'll add more here eventually. I promise.
}
\examples{

library(ggplot2)

example_plot() + theme_steve()

}
