% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_steve.R
\name{theme_steve}
\alias{theme_steve}
\title{Steve's Preferred \pkg{ggplot2} Themes and Assorted Stuff}
\usage{
theme_steve(style = "web", ...)
}
\arguments{
\item{style}{various styles/adjustments to make to the base theme}

\item{...}{optional stuff, though you don't need to think much about this.}
}
\description{
\code{theme_steve()} is my default theme framework for graphs I
make with \pkg{ggplot2}. It starts \code{theme_bw()}, which is available in
\pkg{ggplot2}, but adjusts the margins and axes a bit to my liking. The end
result is, I think, a lovely template for graphs I make in R.
}
\details{
The best use of this function may involve the fonts you have installed
on your system. The user should experiment with various options to see what they
like.
}
\examples{

library(ggplot2)

example_plot() + theme_steve()

}
